+++ clients/admin-ui/src/features/datastore-connections/system_portal_config/forms/ConnectorParameters.tsx
@@ -6,6 +6,7 @@ import {
   CreateSaasConnectionConfig,
   useCreateSassConnectionConfigMutation,
   useGetConnectionConfigDatasetConfigsQuery,
+  useLazyGetAuthorizationUrlQuery,
 } from "datastore-connections/datastore-connection.slice";
 import { useDatasetConfigField } from "datastore-connections/system_portal_config/forms/fields/DatasetConfigField/DatasetConfigField";
 import {
@@ -196,6 +197,7 @@ export const useConnectorForm = ({
   const dispatch = useAppDispatch();
 
   const [isSubmitting, setIsSubmitting] = useState(false);
+  const [isAuthorizing, setIsAuthorizing] = useState(false);
 
   const {
     dropdownOptions: datasetDropdownOptions,
@@ -207,6 +209,7 @@ export const useConnectorForm = ({
   });
 
   const [createSassConnectionConfig] = useCreateSassConnectionConfigMutation();
+  const [getAuthorizationUrl] = useLazyGetAuthorizationUrlQuery();
   const [updateSystemConnectionSecrets] =
     usePatchSystemConnectionSecretsMutation();
   const [patchDatastoreConnection] = usePatchSystemConnectionConfigsMutation();
@@ -317,9 +320,51 @@ export const useConnectorForm = ({
     }
   };
 
+  const handleAuthorization = async (values: ConnectionConfigFormValues) => {
+    const isCreatingConnectionConfig = !connectionConfig;
+    try {
+      setIsAuthorizing(true);
+      if (isCreatingConnectionConfig) {
+        const response = await createSaasConnector(
+          values, // pre-process dataset references
+          secretsSchema!,
+          connectionOption,
+          systemFidesKey,
+          createSassConnectionConfig
+        );
+        // eslint-disable-next-line no-param-reassign
+        connectionConfig = response.connection;
+      } else {
+        await upsertConnectionConfigSecrets(
+          values,
+          secretsSchema!,
+          systemFidesKey,
+          originalSecrets,
+          updateSystemConnectionSecrets
+        );
+      }
+      const authorizationUrl = (await getAuthorizationUrl(
+        connectionConfig!.key
+      ).unwrap()) as string;
+
+      setIsAuthorizing(false);
+
+      // workaround to make sure isAuthorizing is set to false before redirecting
+      setTimeout(() => {
+        window.location.href = authorizationUrl;
+      }, 0);
+    } catch (error) {
+      handleError(error);
+    } finally {
+      setIsAuthorizing(false);
+    }
+  };
+
   return {
     isSubmitting,
+    isAuthorizing,
     handleSubmit,
+    handleAuthorization,
     datasetDropdownOptions,
     selectedDatasetConfigOption,
     handleDelete,
@@ -359,7 +404,9 @@ export const ConnectorParameters: React.FC<ConnectorParametersProps> = ({
 
   const {
     isSubmitting,
+    isAuthorizing,
     handleSubmit,
+    handleAuthorization,
     datasetDropdownOptions,
     selectedDatasetConfigOption,
     handleDelete,
@@ -404,8 +451,10 @@ export const ConnectorParameters: React.FC<ConnectorParametersProps> = ({
         secretsSchema={secretsSchema}
         defaultValues={defaultValues}
         isSubmitting={isSubmitting}
+        isAuthorizing={isAuthorizing}
         onSaveClick={handleSubmit}
         onTestConnectionClick={handleTestConnectionClick}
+        onAuthorizeConnectionClick={handleAuthorization}
         connectionOption={connectionOption}
         connectionConfig={connectionConfig}
         datasetDropdownOptions={datasetDropdownOptions}